# src/config/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "Expiry Date Manager API",
    "description": "REST APIs for Expiry Date Manager application including Auth, User, and Product management"
  },
  "servers": [
    {
      "url": "http://localhost:5001"
    }
  ],
  "paths": {
...
    "/products": {
      "get": {
        "summary": "Get paginated products",
        "description": "Retrieves products nearing expiry with pagination, search query, category, and date filtering.",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "Page number (default: 1)",
            "schema": { "type": "integer", "default": 1 }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Items per page (max: 20, default: 20)",
            "schema": { "type": "integer", "default": 20 }
          },
          {
            "name": "search",
            "in": "query",
            "description": "Search term matching title or UPC barcode",
            "schema": { "type": "string" }
          },
          {
            "name": "category",
            "in": "query",
            "description": "Filter by product category",
            "schema": { "type": "string" }
          },
          {
            "name": "monthsFilter",
            "in": "query",
            "description": "Expiry timeframe filter ('1', '3', 'soon', 'expired')",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Products retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedProductsResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create new product",
        "description": "Adds a new product item with barcode UPC, title, amount, category, and expiry date.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Product created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation error"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ProductInput": {
        "type": "object",
        "required": ["title", "expiryDate"],
        "properties": {
          "title": { "type": "string", "example": "Organic Whole Milk 1L" },
          "upc": { "type": "string", "example": "012345678905" },
          "amount": { "type": "string", "example": "1 L" },
          "category": { "type": "string", "example": "Dairy" },
          "expiryDate": { "type": "string", "format": "date", "example": "2026-09-01" }
        }
      },
      "Product": {
        "type": "object",
        "properties": {
          "_id": { "type": "string", "example": "65b53e7f41bb335d2903e1b9" },
          "title": { "type": "string", "example": "Organic Whole Milk 1L" },
          "upc": { "type": "string", "example": "012345678905" },
          "amount": { "type": "string", "example": "1 L" },
          "category": { "type": "string", "example": "Dairy" },
          "expiryDate": { "type": "string", "format": "date", "example": "2026-09-01" },
          "userId": { "type": "string", "example": "65b53e7f41bb335d2903e1a0" },
          "createdAt": { "type": "string", "example": "2026-08-18T10:00:00.000Z" }
        }
      },
      "ProductResponse": {
        "type": "object",
        "properties": {
          "message": { "type": "string", "example": "Operation successful" },
          "product": { "$ref": "#/components/schemas/Product" }
        }
      },
      "PaginatedProductsResponse": {
        "type": "object",
        "properties": {
          "status": { "type": "string", "example": "success" },
          "products": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/Product" }
          },
          "page": { "type": "integer", "example": 1 },
          "limit": { "type": "integer", "example": 20 },
          "total": { "type": "integer", "example": 5 },
          "totalPages": { "type": "integer", "example": 1 }
        }
      }
    }
  }
}